Rename K2 precipitation files to prepare for syncing fork with upstream repo.
code_write_k2_precipitation.py renamed to code_write_k2_precipitation_file.py
tool_write_k2_precipitation.py renamed to tool_write_k2_precipitation_file.py
